fix: remove toast and redirectory query hooks
