--- src/hooks/useGetRooms.ts
@@ -1,16 +1,0 @@
-import { getRooms, postSignIn } from '@/apis/queries';
-import { IRoom } from '@/app/_interfaces';
-import { getIsLoggedin } from '@/utils/getIsLoggedin';
-import { useQuery } from '@tanstack/react-query';
-
-export const useGetRooms = () => {
-  return useQuery<IRoom[]>({
-    queryKey: ['get-rooms'],
-    queryFn: getRooms,
-    // TODO: include later
-    // suspense: true,
-    retry: 3,
-    staleTime: 1000 * 5,
-    enabled: getIsLoggedin(),
-  });
-};
